src/delivery/pyproject.toml
[project]
name = "delivery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiomqtt>=1.0.0",
    "mavsdk>=3.10.2",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.3",
]
